Sidestep possible Rakudo bug with num64 natives in smartmatch
The above is just a rough guess at what might be the underlying issue. The
smartmatch operator was failing for

	$tag_result ~~ $payload[$idx]

But passing for

	$tag_result.perl.EVAL ~~ $payload[$idx]

When we discovered that it works in numeric context as well, this seems to
close the issue WRT to our purposes -- it demonstrates that the numeric value
roundtrips and would Be What We Meant when we encoded the bugger.