Another workaround for the Num vs Rat
This time we follow the behavior of other parsers by casting to Str and
then to Rat. This might feel counterintuitive or wrong to you -- it does
a bit to me as well.

However, until this is demonstrated to be an issue in the wild, I am
going to punt the issue until I can discuss it in person with some other
Perl 6 hackers.

The other option, in my mind, would be patching the JSON parser to
accept an adverb which would trigger Num-ifying instead of Rat-ing.